Hotfix basecase rolls
