+++ helpers/embed.js
@@ -1,4 +1,8 @@
-const {ActionRowBuilder, ButtonBuilder} = require('discord.js');
+const {ActionRowBuilder, ButtonBuilder, EmbedBuilder} = require('discord.js');
+const urljoin = require('urljoin');
+
+const {albumURL} = require('../config/config.js');
+const {rarities} = require('./constants.js');
 
 function newPanel(){
     const row = new ActionRowBuilder()
@@ -17,4 +21,46 @@ function newPanel(){
     return row;
 }
 
-module.exports = {newPanel};
+
+async function sendCardEmbed(message, cards){
+    const pages = cards.map(c =>{
+        const photopath = urljoin(albumURL,`${c.URLimagen}.png`);
+        const embed = new EmbedBuilder()
+            .setTitle("Informacion de carta")
+            .setColor(0x31593B)
+            .setImage(photopath)
+            .addFields(
+                { name: `Carta`, value:c.nombre},
+                { name: `Serie`, value: `${c.serie}  (${c.numero})`},
+                { name: `Rareza`, value: `${c.rareza}  (${rarities[c.rareza]})`},
+            );
+        return embed;
+    });
+
+    let currentPage = 0;
+    const buttonPanel = newPanel();         
+    const msg = await message.channel.send({ embeds: [pages[currentPage]], components: [buttonPanel] });
+
+    const filter = i => i.customId === 'previous_button' || i.customId === 'next_button';
+    const collector = msg.createMessageComponentCollector({ filter, time: 60000 });
+
+    collector.on('collect', async interaction => {
+        if (interaction.customId === 'previous_button') {
+            currentPage = Math.max(0, currentPage - 1);
+        } else if (interaction.customId === 'next_button') {
+            currentPage = Math.min(pages.length - 1, currentPage + 1);
+        }
+
+        // Update button states based on current page index
+        buttonPanel.components[0].setDisabled(currentPage === 0); // Disable "previous" button on page 0
+        buttonPanel.components[1].setDisabled(currentPage === pages.length - 1); // Disable "next" button on last page
+
+        await interaction.update({ embeds: [pages[currentPage]], components: [buttonPanel] });
+    });
+
+    collector.on('end', async () => {
+        await msg.edit({ components: [] });
+    });
+}
+
+module.exports = {newPanel,sendCardEmbed};
