+++ helpers/index.js
@@ -1,9 +1,10 @@
-const {newPanel} = require('./embed.js');
+const {newPanel,sendCardEmbed} = require('./embed.js');
 const {parseCartas} = require('./parse.js');
 const {rarities} = require('./constants.js');
 
 module.exports = {
     newPanel,
     rarities,
-    parseCartas
+    parseCartas,
+    sendCardEmbed
 };
