+++ index.js
@@ -1,7 +1,6 @@
 const { Client, Events,EmbedBuilder, AttachmentBuilder} = require('discord.js');
 const cron = require('cron');
 const client = new Client({ intents: [37633] });
-const urljoin = require('urljoin');
 const {
     incrementElement,
     roll,
@@ -17,7 +16,7 @@ const {
     help
 } = require("./commands/commands.js");
 const {models} = require("./database.js");
-const { newPanel, rarities } = require('./helpers');
+const { newPanel, rarities, sendCardEmbed} = require('./helpers');
 const config = require('./config/config.js');
 const {token,prefix,albumURL} = config;
 
@@ -73,51 +72,16 @@ client.on('messageCreate', async message => {
                 responses = responses.concat(await repeats(player, message));
                 break;
             case 'roll':
-
                 //Logic that handles the database update and returns the cards that were rolled 
                 cards = await roll(player);
-
-                //Create the embeds for the cards
-                const pages = cards.map(c =>{
-                    const photopath = urljoin(albumURL,`${c.URLimagen}.png`);
-                    const embed = new EmbedBuilder()
-                        .setTitle("Informacion de carta")
-                        .setColor(0x31593B)
-                        .setImage(photopath)
-                        .addFields(
-                            { name: `Carta`, value:c.nombre},
-                            { name: `Serie`, value: `${c.serie}  (${c.numero})`},
-                            { name: `Rareza`, value: `${c.rareza}  (${rarities[c.rareza]})`},
-                        );
-                    return embed;
-                });
-
-                let currentPage = 0;
-                //I dont understand javascript controllers
-                const buttonPanel = newPanel();         
-                const msg = await message.channel.send({ embeds: [pages[currentPage]], components: [buttonPanel] });
-        
-                const filter = i => i.customId === 'previous_button' || i.customId === 'next_button';
-                const collector = msg.createMessageComponentCollector({ filter, time: 60000 });
-        
-                collector.on('collect', async interaction => {
-                    if (interaction.customId === 'previous_button') {
-                        currentPage = Math.max(0, currentPage - 1);
-                    } else if (interaction.customId === 'next_button') {
-                        currentPage = Math.min(pages.length - 1, currentPage + 1);
-                    }
-        
-                    // Update button states based on current page index
-                    buttonPanel.components[0].setDisabled(currentPage === 0); // Disable "previous" button on page 0
-                    buttonPanel.components[1].setDisabled(currentPage === pages.length - 1); // Disable "next" button on last page
-        
-                    await interaction.update({ embeds: [pages[currentPage]], components: [buttonPanel] });
-                });
-        
-                collector.on('end', async () => {
-                    await msg.edit({ components: [] });
-                });
-                
+                if (!cards.length) {
+                    // Rolls will only return an empty list when the player has 
+                    // no rolls
+                    responses.push("No tienes rolls");
+                    break;
+                }
+                // Create and send the embeds for the cards
+                await sendCardEmbed(message,cards);
                 break;    
             case 'trade':
                 responses = responses.concat("Pato");
